fix(CMakeLists): link SDL2 libraries when USE_SDL2
In bd4545c ('add options to select backend') the default option is to
use the framebuffer. When `USE_SDL2=ON` or `USE_DRM=ON` are passed to
CMake, the framebuffer option must be turned off. This is fixed with the
present commit.

Also, in order for these options to apply to lvgl,
`add_subdirectory(lvgl ...)` needs to be moved after the option setting
code.

Signed-off-by: Jakob Kastelic <jkastelic@thinksrs.com>